IMPLEMENTED  GENERATE RETURNSLIP AND CLAIMING RETURNSLIP

+++ new-sale-api/src/main/java/com/otsi/retail/newSale/vo/ListOfReturnSlipsVo.java
@@ -18,7 +18,7 @@ import lombok.Data;
 @Data
 public class ListOfReturnSlipsVo {
 	
-	private int rsId;
+	private Long rsId;
 
 	private LocalDate dateFrom;
